requirements/common.txt
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
uvloop==0.22.1
pydub==0.25.1
librosa==0.11.0
zhon==2.1.1
loguru==0.7.3
ruamel.yaml==0.19.1
fastapi==0.128.0
schedule==1.2.2
pytz==2025.2
jinja2==3.1.6
ray==2.53.0
tqdm==4.67.1
omegaconf==2.3.0
inflect==7.5.0
hydra-core==1.3.2
rich==14.2.0
gdown==5.2.1
matplotlib==3.10.8
wget==3.2
seg2stream==1.0.8
conformer==0.3.2
openai-whisper==20250625
lightning==2.6.0
diffusers==0.36.0
x_transformers==2.14.2
wetext==0.1.2
#############################
torch==2.7.1
torchaudio==2.7.1
onnxruntime-gpu==1.23.2
onnx==1.19.1
onnxsim==0.4.36
onnxslim==0.1.82
tensorrt-cu12==10.9.0.34
numpy==2.3.5